Refactor code structure and fix styling issues in canvas.js and drawer.js
