--- config.js
@@ -1,1 +1,0 @@
-const modes = ['draw', 'select', 'delete', 'drag']
