+++ config/config.js
@@ -1,0 +1,4 @@
+const modes = ['draw', 'select', 'delete', 'drag']
+export var gridSize = 20
+export var defaultPointSize = 0.5
+
